+++ hosts/upsilon/home.nix
@@ -1,5 +1,0 @@
-{ keys, ... }:
-
-{
-  programs.git.extraConfig.user.signingKey = keys.users."michael.hoang_upsilon";
-}