+++ keys.nix
@@ -6,8 +6,6 @@
       "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAILzD/3cOhlqe8NVEruSUnPSnG1GbmX8SgTbVGLFHMa7g";
     enzime_sigma =
       "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIHwM0sVLdolDXb7LyKMT6Jfym0dP1NSgJClHcIOga+fa";
-    "michael.hoang_upsilon" =
-      "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIIItcLLX1d9G1+SdjaMkAAmlkQ34L8OUN7KwWQkGblK1";
   };
 
   hosts = {
@@ -19,8 +17,6 @@
       "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIMxOi/S1TLBg8/ZRX5XfCTlM8A+I0q0pQksrxtfjdYFP";
     sigma =
       "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIDxRoznXzz/T6s5UeHG1uoHCXGfXSpy27eTEzC0/EUW+";
-    upsilon =
-      "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIKg43qekSo5Cv4eEFJTLXTuXt8w2rsFtY1QLuB5Qlofp";
   };
 
   signing = {